refactor(asset): mostly modularise asset/util (formerly asset-io)
